# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ArchSmart Lite API",
    "version": "1.0.0",
    "description": "Dokumentasi REST API untuk platform edukasi ArchSmart Lite. Menggunakan Supabase sebagai backend data."
  },
  "servers": [
    {
      "url": "{serverUrl}",
      "description": "Current Environment URL"
    }
  ],
  "tags": [
    {
      "name": "Courses",
      "description": "Endpoint untuk data materi pembelajaran"
    },
    {
      "name": "Bookmarks",
      "description": "Manajemen materi yang disimpan user"
    }
  ],
  "paths": {
    "/api/courses": {
      "get": {
        "tags": ["Courses"],
        "summary": "Ambil semua daftar kursus",
        "description": "Mengembalikan daftar semua kursus yang tersedia di database.",
        "responses": {
          "200": {
            "description": "Berhasil mengambil data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Course" }
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{id}": {
      "get": {
        "tags": ["Courses"],
        "summary": "Ambil detail satu kursus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Detail kursus ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Course" }
              }
            }
          },
          "404": { "description": "Kursus tidak ditemukan" }
        }
      }
    },
    "/api/bookmarks": {
      "get": {
        "tags": ["Bookmarks"],
        "summary": "Ambil daftar bookmark user",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "description": "UUID User (dari tabel auth.users)",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar kursus yang dibookmark",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Course" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Bookmarks"],
        "summary": "Simpan bookmark baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": { "type": "string", "format": "uuid" },
                  "course_id": { "type": "string", "format": "uuid" }
                },
                "required": ["user_id", "course_id"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Bookmark berhasil disimpan" },
          "400": { "description": "Data user_id atau course_id tidak valid" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Course": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "category": { "type": "string" },
          "description": { "type": "string" },
          "image_url": { "type": "string", "format": "uri" },
          "price": { "type": "string" },
          "duration": { "type": "string" },
          "modules_count": { "type": "integer" },
          "rating": { "type": "number" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}